78 use cfg_file_paths, resolve_paths->resolve_cfg_file, extract fs.ch
